+++ src/Server.hs
@@ -4,8 +4,9 @@ module Server (runServer) where
 
 import Handle
 import Database
-import Network.Gopher (runGopher, defaultConfig, cServerPort, cLogHandler, GopherLogHandler, fromGopherLogStr)
+import Network.Gopher (runGopher, defaultConfig, cServerPort, cLogHandler, GopherLogHandler, fromGopherLogStr, GopherConfig (cServerName))
 import Config
+import Data.Text.Encoding (encodeUtf8)
 
 logHandler :: GopherLogHandler
 logHandler level str = do
@@ -16,4 +17,4 @@ runServer :: Config -> IO ()
 runServer config = do
   _ <- initializeDatabase config
   putStrLn "Starting spacecookie server"
-  runGopher (defaultConfig { cServerPort = 7000, cLogHandler = Just logHandler }) (handler config)
+  runGopher (defaultConfig { cServerName = encodeUtf8 config.spacecookie.name, cServerPort = fromIntegral config.spacecookie.port, cLogHandler = Just logHandler }) (handler config)