+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
   * Configurable maximum post length
   * Configuration for client sent an empty query language
+  * Config options for spacecookie server
 
 ### Fixed